prevent relogins but in a sketch way
i literally have no idea how to clear the session and since this is a POC no one will actually use, we can just check if the raw JSON is null. if so, then they never logged in through the IdP login page. they bypassed it as an SP was reinitating with an existing session. so we store if they were denied / allowed and then reuse that answer when the JSON is null. shibboleth is so freaking confusing and i have a headache. i am supposed to be doing things that lower my blood pressure, mot raise it. i know this is sketch but trying to find the "shibboleth" way was not successful after multiple hours. i don't want to bug the duke iam team again. i am losing it!